(~ re lex) Inline the lexing of inline comments. Meta.
Note that I've opted to *include* the comment-delimiter in the token in this case. This is still differentiable upstream by the presence/lack of COMMENT_OPEN or COMMENT_CLOSE delimiters around the COMMENT chunk. I'm doing this in preperation for adding Vim-style `"` comments as well; as well as to match how I suspect StringComplex lexing is going to work out ...