(re meta) Reimplement Menhir type-annotation machinery due to fpottier/menhir#24
This was a mess, holy lord. Details in the linked GitLab issue; but basically, Menhir's output is inexplicably broken UTF-8, which JavaScript tooling is "fixing" by substituting U+FFFD / UTF-8 \xEF\xBF\xBD. That, of course, completely breaks the automaton.

Solution? Re-write all the manipulation using Buffers, at least for now. Now no part of the input-file gets round-tripped through String; and is treated as opaque data; only the search-terms are required to be valid UTF-8. (This is going to be a problem if the ppx tools ever support GADTs, though ... because I'll need regexen to match that, as I briefly had previously.)

Luckily, this hack works. For now.

/see https://gitlab.inria.fr/fpottier/menhir/issues/24