(NF re MLAPI err) Revamping LexErrors, raising error-handling into parser
This is why I undertook this project! So, in the category of "lessons Elliott's learned about parser-design", it turns out exceptions are a poor error-recovery mechanism in the context of a recoverable or incremental parser.

I'm still leaving in the 'legacy'(?) ability to dump errors out to exceptional control-flow at the boundaries of the lexer (cf. `Lexer.next`, `Lexer.next_exn`); but now, I'm hoping to ensure a majority of tokenization errors "bubble up" to the Menhir automaton, so they can be queried (and rolled back) thru the incremental interface.

Incomplete work, couple tests failing, only one particular type of lexer-error has been converted so far, yadda yadda yadda.