(!! re AST JSintf MLAPI) Support subexpressions in the AST and interfaces
This is the substantial portion of support for piping/subexpressions; and involves refactoring ...

1. our datastructure; to support the representation of "holes" in the tree for unevaluated
   subexpressions,
2. and the interfaces; to *interact* with those "holes."

Notably:

 - Several AST nodes are now wrapped in `AST.or_subexpr`; these can either be a `Literal`
   (equivalent to the previous behaviour), or a `Sub`, pending evaluation;
 - the ML and JavaScript interfaces to `Expression` now return these values, whether `Literal` or
   `Sub`, leaving the decision of how to proceed (further evaluating the subexpression or not) up to
   the caller;
 - and a new JavaScript convience feature, in the form of `ExpressionEval` and the `eval*` functions
   on `Expression`, has been provided.

The new `eval` interfaces provide a way to quickly produce simple `string` values for AST nodes,
assuming you can provide a simple function that will map any subexpression itself to a string — it's
basically an inversion-of-control interface, in which the AST datastructure can invoke your
"evaluator" function as-necessary.

The final step, actually producing subexpressions in the parser, is next!

(Finally, some of the above necessitated AST-duplication, so there's now a bevy of `copy` functions
available.)