(re parse MLAPI) Build, and store, arguments as a reversed list
This is a bit subtle; but Menhir (like its ancestor, yacc) constructs a pushdown automaton, i.e. a *left-recursive* parser. Generally, this means it's a good idea to build correspondingly left-recursive grammars.

With this commit, I've propagated this practice through the `arguments` nonterminals.

By consistently retaining the arguments-list in reversed form *throughout* the parsing pipeline, I've removed a ton of cruft, memory-wasting duplication, and reversing and re-reversing of lists. Phew.

In the process, I've done away with awkward `array`-manipulation; although there's a forseeable downside in terms of debugability: I doubt I'm going to be able to train users to enable `-bs-g`.