(!! new doc) Generating documentation from the TypeScript interface
I'm going to be moving towards the JavaScript-interface as a 'single source of truth' for documentation, even though that's going to dislocate any in-depth documentation I write away from the implementation it refrences.

One thing I'd *really* like, is a direct hyperlink from the JS-interface docs for a function, to the corresponding ML-docs for that function. Maybe in the future.