(!! NF rm) Remove CommonJS support entirely
Publishing a dual-module-system project is a lot of work, a lot of fragility, and simply not important for this project's "userbase" (read: one dude. :P)

There's still, technically, a `.cjs` build to `dist/excmd.cjs`; but I'm not gonna put a ton of work into supporting it. It's 2021, `.mjs` support is LTS in Node.js v14+ (and even in v12+ with `--experimental-modules`); let's move forward.

Notably, there's a regression in Node itself I'm running into causing an `ERR_PACKAGE_PATH_NOT_EXPORTED`, so this commit doesn't yet build - fixes incoming:

https://github.com/babel/babel/issues/11216